♻️ refactor: type 파일 정리
