+++ feed-crawler/src/repository/feed.repository.ts
@@ -1,12 +1,11 @@
 import { inject, injectable } from 'tsyringe';
 
 import { redisConstant } from '@common/constant';
+import { DatabaseConnection } from '@common/database-connection';
+import { DEPENDENCY_SYMBOLS } from '@common/dependency-symbols';
 import logger from '@common/logger';
 import { RedisConnection } from '@common/redis-access';
 import { FeedDetail } from '@common/types';
-
-import { DatabaseConnection } from '@app-types/database-connection';
-import { DEPENDENCY_SYMBOLS } from '@app-types/dependency-symbols';
 
 @injectable()
 export class FeedRepository {
