+++ feed-crawler/src/repository/rss.repository.ts
@@ -1,9 +1,8 @@
 import { inject, injectable } from 'tsyringe';
 
+import { DatabaseConnection } from '@common/database-connection';
+import { DEPENDENCY_SYMBOLS } from '@common/dependency-symbols';
 import { RssObj } from '@common/types';
-
-import { DatabaseConnection } from '@app-types/database-connection';
-import { DEPENDENCY_SYMBOLS } from '@app-types/dependency-symbols';
 
 @injectable()
 export class RssRepository {
