+++ feed-crawler/src/repository/tag-map.repository.ts
@@ -1,7 +1,7 @@
 import { inject, injectable } from 'tsyringe';
 
-import { DatabaseConnection } from '@app-types/database-connection';
-import { DEPENDENCY_SYMBOLS } from '@app-types/dependency-symbols';
+import { DatabaseConnection } from '@common/database-connection';
+import { DEPENDENCY_SYMBOLS } from '@common/dependency-symbols';
 
 @injectable()
 export class TagMapRepository {
